requirements.txt
flask==2.3.3
flask-cors==3.0.10
openai==0.27.0
nltk==3.8.1
rich==13.7.0
python-dotenv==1.0.0
httpx==0.23.3